Add API folder
