+++ API/Food recognition/swagger.yaml
@@ -1,0 +1,156 @@
+openapi: 3.0.0
+info:
+  title: 영칼로리 API
+  description: |-
+    **OSS Team Project**
+    
+    **장정우**
+  version: 1.0.0
+servers:
+  - url: http://localhost:5000
+    description: Local server
+paths:
+  /quantity:
+    post:
+      tags:
+        - Food Recognition    
+      summary: 음식량 인식
+      requestBody:
+        required: true
+        content:
+          application/json:
+            schema:
+              type: object
+              properties:
+                image_path:
+                  type: string
+                  description: 인식할 이미지 경로
+                  example: "./image.jpg"
+      responses:
+        '200':
+          description: 성공
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  quantity:
+                    type: number
+                    description: 추정된 음식량
+                    example: 1.25
+        '400':
+          description: 잘못된 입력
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  error:
+                    type: string
+                    example: image_path is required
+        '500':
+          description: 서버 에러
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  error:
+                    type: string
+                    example: An error occurred
+
+  /detect:
+    post:
+      summary: 음식 인식
+      tags:
+        - Food Recognition
+      requestBody:
+        required: true
+        content:
+          application/json:
+            schema:
+              type: object
+              properties:
+                image_path:
+                  type: string
+                  description: 인식할 이미지 경로
+                  example: "./image.jpg"
+      responses:
+        '200':
+          description: 성공
+          content:
+            application/json:
+              schema:
+                type: array
+                items:
+                  type: object
+                  properties:
+                    food:
+                      type: string
+                      description: Name of food
+                      example: Apple
+                    calories:
+                      type: string
+                      description: Calories
+                      example: 52.00 kcal
+                    carbs:
+                      type: string
+                      description: Carb
+                      example: 14.00 g
+                    protein:
+                      type: string
+                      description: Protein
+                      example: 0.30 g
+                    fat:
+                      type: string
+                      description: Fat
+                      example: 0.20 g
+        '400':
+          description: 잘못된 입력
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  error:
+                    type: string
+                    example: image_path is required
+        '500':
+          description: 서버 에러
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  error:
+                    type: string
+                    example: An error occurred
+
+  /test:
+    get:
+      summary: API 테스트
+      tags:
+        - Test
+      responses:
+        '200':
+          description: API 테스트 성공
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  message:
+                    type: string
+                    description: 성공 메시지
+                    example: API is working correctly
+        '500':
+          description: 서버 에러
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  error:
+                    type: string
+                    description: 오류 메시지
+                    example: An error occurred
